[IDLE-78] androidXFragment 의존성 추가 및 Auth 모듈 ViewModel 추가
