+++ feature/auth/src/main/java/com/idle/auth/AuthFragment.kt
@@ -1,0 +1,73 @@
+package com.idle.auth
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.ComposeView
+import androidx.compose.ui.unit.dp
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.viewModels
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+internal class AuthFragment : Fragment() {
+
+    private lateinit var composeView: ComposeView
+    val viewModel: AuthViewModel by viewModels()
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ): View {
+        return ComposeView(requireContext()).also {
+            composeView = it
+        }
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        composeView.setContent {
+            AuthScreen()
+        }
+    }
+}
+
+@Composable
+internal fun AuthScreen() {
+    Column(
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.spacedBy(20.dp, Alignment.CenterVertically),
+        modifier = Modifier.fillMaxSize()
+            .padding(horizontal = 20.dp),
+    ) {
+        Spacer(
+            modifier = Modifier.height(300.dp)
+                .fillMaxWidth()
+                .background(Color.Black)
+        )
+
+        Button(onClick = {}) {
+            Text(text = "요양 보호사로 시작하기")
+        }
+
+        Button(onClick = {}) {
+            Text(text = "센터 관리자로 시작하기")
+        }
+    }
+}
