--- feature/auth/src/main/java/com/idle/auth/AuthRoute.kt
@@ -1,48 +1,0 @@
-package com.idle.auth
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.ComposeView
-import androidx.fragment.app.Fragment
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-internal class AuthFragment : Fragment() {
-
-    private lateinit var composeView: ComposeView
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?,
-    ): View {
-        return ComposeView(requireContext()).also {
-            composeView = it
-        }
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        composeView.setContent {
-            AuthScreen()
-        }
-    }
-}
-
-@Composable
-internal fun AuthScreen() {
-    Box(modifier = Modifier.fillMaxSize()) {
-        Text(
-            text = "init project",
-            modifier = Modifier.align(Alignment.Center),
-        )
-    }
-}
