+++ feature/auth/src/main/java/com/idle/auth/AuthViewModel.kt
@@ -1,0 +1,8 @@
+package com.idle.auth
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class AuthViewModel @Inject constructor() : ViewModel() {}
